Update AllocationAnalysis
Found and fixed an issue with the number of records it was pulling (it wasn't pulling all the pending claimants, only 1690ish total). Also fixed the Time_Since_SA case statement to reflect correct statement for claimants with a NULL allocation (addressed this by changing #AllData and the case statement).