Create 5Box
This procedure calculates the date a lien was created and the date it was pushed to final (or closed, if it went straight to closed and skipped a final stage). Then, it creates a record for each lien showing pending for each month in between the created date and the finalized date. This was created in order to "force" a row for each lien for every month that the lien existed but was not final. With this data, you can take a proportion of how many liens were final vs pending at any given month.